requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.4
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
boto3==1.28.23
botocore==1.31.54
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.10.0
Cython==3.0.0
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
easyocr==1.7.1
et-xmlfile==1.1.0
fastapi==0.110.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==2.10
imageio==2.34.0
inference==0.9.17
inference-cli==0.9.17
inference-sdk==0.9.17
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.15.1
openai==1.17.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.3.0
pluggy==1.4.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
protobuf==5.26.1
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pybase64==1.3.2
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySide2==5.13.2
pytesseract==0.3.10
pytest==8.1.1
pytest-asyncio==0.21.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.27
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.6.2
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.1
shiboken2==5.13.2
six==1.16.0
skypilot==0.4.1
sniffio==1.3.1
starlette==0.37.2
structlog==24.1.0
supervision==0.19.0
sympy==1.12
tabulate==0.9.0
tesseract==0.1.3
threadpoolctl==3.4.0
tifffile==2024.2.12
time-machine==2.14.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
